Simplify thumbnail axis overlay and move helpers into review_gui
Keep isaac_sim_debug_draw focused on Kit debug draw, and trim duplicate
prim/AABB/PNG plumbing in thumbnail capture.
