+++ isaaclab_arena/utils/isaac_sim_debug_draw.py
@@ -9,9 +9,6 @@ from __future__ import annotations
 
 # Default color for all bounding boxes (bright green)
 DEFAULT_COLOR = (0.0, 1.0, 0.0, 1.0)
-AXIS_X_COLOR = (1.0, 0.0, 0.0, 1.0)
-AXIS_Y_COLOR = (0.0, 1.0, 0.0, 1.0)
-AXIS_Z_COLOR = (0.0, 0.0, 1.0, 1.0)
 
 
 class IsaacSimDebugDraw:
@@ -144,105 +141,3 @@ class IsaacSimDebugDraw:
         widths = [thickness] * len(edges)
 
         self._draw.draw_lines(start_points, end_points, colors_list, widths)
-
-
-def local_xyz_axis_segments(
-    origin_xyz: tuple[float, float, float],
-    *,
-    x_dir: tuple[float, float, float],
-    y_dir: tuple[float, float, float],
-    z_dir: tuple[float, float, float],
-    length_m: float,
-) -> tuple[
-    list[tuple[float, float, float]],
-    list[tuple[float, float, float]],
-    list[tuple[float, float, float, float]],
-]:
-    """Return start/end points and RGBA colors for local +X/+Y/+Z axis lines.
-
-    Args:
-        origin_xyz: World-space origin of the axes.
-        x_dir: Direction of local +X in world space (normalized internally).
-        y_dir: Direction of local +Y in world space (normalized internally).
-        z_dir: Direction of local +Z in world space (normalized internally).
-        length_m: Axis length in meters.
-
-    Returns:
-        ``(start_points, end_points, colors)`` suitable for ``draw_lines``.
-    """
-    assert length_m > 0.0, f"length_m must be positive, got {length_m}"
-    axes = (
-        (_unit(x_dir), AXIS_X_COLOR),
-        (_unit(y_dir), AXIS_Y_COLOR),
-        (_unit(z_dir), AXIS_Z_COLOR),
-    )
-    start_points: list[tuple[float, float, float]] = []
-    end_points: list[tuple[float, float, float]] = []
-    colors: list[tuple[float, float, float, float]] = []
-    for direction, color in axes:
-        if direction is None:
-            continue
-        start_points.append(origin_xyz)
-        end_points.append((
-            origin_xyz[0] + direction[0] * length_m,
-            origin_xyz[1] + direction[1] * length_m,
-            origin_xyz[2] + direction[2] * length_m,
-        ))
-        colors.append(color)
-    return start_points, end_points, colors
-
-
-def axis_length_from_extents(
-    extents_xyz: tuple[float, float, float], *, fraction: float = 0.35, min_m: float = 0.05
-) -> float:
-    """Pick a visible axis length from AABB extents.
-
-    Args:
-        extents_xyz: Axis-aligned size ``(x, y, z)`` in meters.
-        fraction: Fraction of the largest extent to use.
-        min_m: Minimum length so tiny assets still show axes.
-
-    Returns:
-        Axis length in meters.
-    """
-    assert fraction > 0.0, f"fraction must be positive, got {fraction}"
-    assert min_m > 0.0, f"min_m must be positive, got {min_m}"
-    return max(min_m, fraction * max(extents_xyz))
-
-
-def overlay_rgb_segments_on_png(
-    png_bytes: bytes,
-    segments: list[tuple[tuple[int, int], tuple[int, int], tuple[int, int, int]]],
-    *,
-    width_px: int = 4,
-) -> bytes:
-    """Draw opaque RGB line segments on top of a PNG and return new PNG bytes.
-
-    Args:
-        png_bytes: Source snapshot PNG.
-        segments: ``(start_uv, end_uv, rgb)`` pixel segments.
-        width_px: Stroke width in pixels.
-
-    Returns:
-        PNG bytes with the segments composited on top.
-    """
-    import io
-
-    from PIL import Image, ImageDraw
-
-    with Image.open(io.BytesIO(png_bytes)) as image:
-        canvas = image.convert("RGB")
-        draw = ImageDraw.Draw(canvas)
-        for start, end, rgb in segments:
-            draw.line([start, end], fill=rgb, width=width_px)
-        out = io.BytesIO()
-        canvas.save(out, format="PNG")
-        return out.getvalue()
-
-
-def _unit(direction: tuple[float, float, float]) -> tuple[float, float, float] | None:
-    """Normalize a 3D direction, or None when the vector is degenerate."""
-    length = (direction[0] ** 2 + direction[1] ** 2 + direction[2] ** 2) ** 0.5
-    if length < 1e-12:
-        return None
-    return (direction[0] / length, direction[1] / length, direction[2] / length)
